requirements.txt
alembic==1.13.1
aniso8601==9.0.1
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2        
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8     
distlib==0.3.8
filelock==3.14.0
Flask==3.0.3
flask-marshmallow==1.2.1 
Flask-Migrate==4.0.7     
Flask-RESTful==0.3.10    
Flask-SQLAlchemy==3.1.1  
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0      
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5        
marshmallow==3.21.2      
packaging==24.0
platformdirs==4.2.1      
psycopg==3.1.19
psycopg-binary==3.1.19   
psycopg2-binary==2.9.9   
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0        
python-dotenv==1.0.1     
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.30       
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
virtualenv==20.26.2
Werkzeug==3.0.3